feat(dev): cambio en el tipo de dato de los repositorios xd

+++ src/main/java/pe/edu/upc/MonolithFoodApplication/repositories/EatRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.MonolithFoodApplication.entities.EatEntity;
 
 @Repository
-public interface EatRepository extends JpaRepository<EatEntity, Integer> {
+public interface EatRepository extends JpaRepository<EatEntity, Long> {
     
 }
